drastically simplify; invert getattr/getitem; down to 54 fails on 3.5
