+++ attmap/attmap_echo.py
@@ -1,14 +1,40 @@
 """ AttMap that echoes an unset key/attr """
 
 from .ordpathex_attmap import OrdPathExAttMap
-from ._mixins import *
 
 
-class AttMapEcho(EchoMixin, OrdPathExAttMap):
+class AttMapEcho(OrdPathExAttMap):
     """ An AttMap that returns key/attr if it has no set value. """
 
-    def __contains__(self, item):
-        return item in self.__dict__
+    def __getattr__(self, item, default=None):
+        """
+        Fetch the value associated with the provided identifier.
+
+        :param int | str item: identifier for value to fetch
+        :return object: whatever value corresponds to the requested key/item
+        :raises AttributeError: if the requested item has not been set,
+            no default value is provided, and this instance is not configured
+            to return the requested key/item itself when it's missing; also,
+            if the requested item is unmapped and appears to be protected,
+            i.e. by flanking double underscores, then raise AttributeError
+            anyway. More specifically, respect attribute naming that appears
+            to be indicative of the intent of protection.
+        """
+        try:
+            return super(self.__class__, self).__getattr__(item, default)
+        except (AttributeError, TypeError):
+            # If not, triage and cope accordingly.
+            if item.startswith("__") and item.endswith("__"):
+                # Accommodate security-through-obscurity approach used by some libraries.
+                error_reason = "Protected-looking attribute: {}".format(item)
+                raise AttributeError(error_reason)
+            if default is not None:
+                # For compatibility with ordinary getattr() call, allow default value.
+                return default
+            return item
+
+    #def __contains__(self, item):
+    #    return item in self.__dict__
 
     @property
     def _lower_type_bound(self):
